processing_edges/rosbag_to_nuscenes/requirements_lock.linux.x86_64.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --allow-unsafe --output-file=requirements_lock.linux.x86_64.txt requirements.txt
#
contourpy==1.1.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dpkt==1.9.8
    # via velodyne-decoder
fonttools==4.57.0
    # via matplotlib
importlib-resources==6.4.5
    # via
    #   matplotlib
    #   velodyne-decoder
kiwisolver==1.4.7
    # via matplotlib
lxml==6.0.0
    # via -r requirements.txt
lz4==4.3.3
    # via rosbags
matplotlib==3.7.5
    # via -r requirements.txt
numpy==1.24.4
    # via
    #   -r requirements.txt
    #   contourpy
    #   matplotlib
    #   opencv-python
    #   opencv-python-headless
    #   rosbags
    #   rosbags-image
    #   scipy
    #   velodyne-decoder
opencv-python==4.12.0.88
    # via -r requirements.txt
opencv-python-headless==4.12.0.88
    # via rosbags-image
packaging==25.0
    # via matplotlib
pillow==10.4.0
    # via matplotlib
pyparsing==3.1.4
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
pyyaml==6.0.2
    # via
    #   -r requirements.txt
    #   velodyne-decoder
rosbags==0.9.23
    # via
    #   -r requirements.txt
    #   rosbags-image
rosbags-image==0.9.2
    # via -r requirements.txt
ruamel-yaml==0.18.14
    # via rosbags
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
scipy==1.10.1
    # via -r requirements.txt
six==1.17.0
    # via python-dateutil
velodyne-decoder==3.1.0
    # via -r requirements.txt
zipp==3.20.2
    # via importlib-resources
zstandard==0.23.0
    # via rosbags
